+++ lib/reactive_forms.dart
@@ -28,4 +28,5 @@ export 'src/widgets/reactive_slider.dart';
 export 'src/widgets/reactive_status_listenable_builder.dart';
 export 'src/widgets/reactive_switch.dart';
 export 'src/widgets/reactive_text_field.dart';
+export 'src/widgets/reactive_type_def.dart';
 export 'src/widgets/reactive_value_listenable_builder.dart';